lint
SCCS-vsn: sys/vax/vax/Locore.c 6.8
SCCS-vsn: sys/vax/vax/crl.c 1.3
SCCS-vsn: sys/vax/vax/ioa.h 1.3
SCCS-vsn: sys/vax/vax/locore.s 6.35
SCCS-vsn: sys/vax/vax/machdep.c 6.26
SCCS-vsn: sys/vax/vax/mem.c 6.4
SCCS-vsn: sys/vax/vax/trap.c 6.8
SCCS-vsn: sys/vax/vax/tu.c 6.7
SCCS-vsn: sys/kern/init_main.c 6.15
SCCS-vsn: sys/kern/kern_exec.c 6.17
SCCS-vsn: sys/kern/kern_fork.c 6.9
SCCS-vsn: sys/kern/kern_prot.c 6.8
SCCS-vsn: sys/kern/kern_sig.c 6.21
SCCS-vsn: sys/kern/kern_time.c 6.9
SCCS-vsn: sys/kern/subr_log.c 6.8
SCCS-vsn: sys/kern/subr_prf.c 6.12
SCCS-vsn: sys/kern/subr_xxx.c 6.7
SCCS-vsn: sys/kern/tty.c 6.26
SCCS-vsn: sys/kern/tty_pty.c 6.20
SCCS-vsn: sys/ufs/ffs/ffs_inode.c 6.19
SCCS-vsn: sys/ufs/ffs/ufs_inode.c 6.19
SCCS-vsn: sys/ufs/lfs/lfs_inode.c 6.19
SCCS-vsn: sys/ufs/ufs/ufs_inode.c 6.19
SCCS-vsn: sys/kern/vfs_lookup.c 6.28
SCCS-vsn: sys/ufs/ffs/ufs_lookup.c 6.28
SCCS-vsn: sys/ufs/ufs/ufs_lookup.c 6.28
SCCS-vsn: sys/kern/vfs_syscalls.c 6.22
SCCS-vsn: sys/ufs/ffs/ffs_vnops.c 6.22
SCCS-vsn: sys/ufs/ffs/ufs_vnops.c 6.22
SCCS-vsn: sys/ufs/lfs/lfs_vnops.c 6.22
SCCS-vsn: sys/ufs/ufs/ufs_vnops.c 6.22
SCCS-vsn: sys/kern/uipc_socket.c 6.23
SCCS-vsn: sys/kern/uipc_socket2.c 6.15
SCCS-vsn: sys/kern/uipc_syscalls.c 6.15
SCCS-vsn: sys/kern/uipc_usrreq.c 6.21
SCCS-vsn: sys/vax/uba/dmf.c 6.15
SCCS-vsn: sys/vax/uba/lp.c 6.8
SCCS-vsn: sys/vax/uba/lpa.c 6.4
SCCS-vsn: sys/vax/uba/ps.c 6.7
SCCS-vsn: sys/vax/uba/rl.c 6.7
SCCS-vsn: sys/vax/uba/tm.c 6.10
SCCS-vsn: sys/vax/uba/tmscp.c 5.4
SCCS-vsn: sys/vax/uba/uba.c 6.10
SCCS-vsn: sys/vax/uba/uda.c 6.19
SCCS-vsn: sys/vax/uba/up.c 6.10
SCCS-vsn: sys/vax/uba/ut.c 6.8
SCCS-vsn: sys/vax/mba/hp.c 6.20
SCCS-vsn: sys/vax/mba/mba.c 6.8
SCCS-vsn: sys/vax/mba/mt.c 6.6
SCCS-vsn: sys/net/if_sl.c 5.3
SCCS-vsn: sys/net/raw_usrreq.c 6.13
SCCS-vsn: sys/netinet/if_ether.c 6.17
SCCS-vsn: sys/netinet/in.c 6.13
SCCS-vsn: sys/netinet/in.h 6.10
SCCS-vsn: sys/netinet/in_pcb.c 6.14
SCCS-vsn: sys/netinet/in_var.h 6.4
SCCS-vsn: sys/netinet/ip_icmp.c 6.16
SCCS-vsn: sys/netinet/ip_input.c 6.20
SCCS-vsn: sys/netinet/ip_output.c 6.13
SCCS-vsn: sys/netinet/raw_ip.c 6.8
SCCS-vsn: sys/netinet/tcp_input.c 6.21
SCCS-vsn: sys/netinet/tcp_output.c 6.14
SCCS-vsn: sys/netinet/tcp_usrreq.c 6.10
SCCS-vsn: sys/netns/ns_error.c 6.7
SCCS-vsn: sys/netns/ns_output.c 6.10
SCCS-vsn: sys/vax/if/if_ddn.c 6.5
SCCS-vsn: sys/vax/if/if_ec.c 6.18
SCCS-vsn: sys/vax/if/if_ex.c 6.14
SCCS-vsn: sys/vax/if/if_hdh.c 6.6
SCCS-vsn: sys/vax/if/if_hy.c 6.9
SCCS-vsn: sys/vax/if/if_uba.c 6.11
SCCS-vsn: sys/vax/if/if_uba.h 6.6
SCCS-vsn: sys/vax/if/if_vv.c 6.18

+++ usr/src/sys/kern/kern_sig.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)kern_sig.c	6.20 (Berkeley) %G%
+ *	@(#)kern_sig.c	6.21 (Berkeley) %G%
  */
 
 #include "../machine/reg.h"
@@ -533,7 +533,7 @@ issig()
 			sigbits &= ~stopsigmask;
 		if (sigbits == 0)
 			break;
-		sig = ffs(sigbits);
+		sig = ffs((long)sigbits);
 		mask = sigmask(sig);
 		p->p_sig &= ~mask;		/* take the signal! */
 		p->p_cursig = sig;
@@ -789,17 +789,17 @@ core()
 	u.u_error = rdwri(UIO_WRITE, ip,
 	    (caddr_t)&u,
 	    ctob(UPAGES),
-	    0, 1, (int *)0);
+	    (off_t)0, 1, (int *)0);
 	if (u.u_error == 0)
 		u.u_error = rdwri(UIO_WRITE, ip,
 		    (caddr_t)ctob(dptov(u.u_procp, 0)),
-		    ctob(u.u_dsize),
-		    ctob(UPAGES), 0, (int *)0);
+		    (int)ctob(u.u_dsize),
+		    (off_t)ctob(UPAGES), 0, (int *)0);
 	if (u.u_error == 0)
 		u.u_error = rdwri(UIO_WRITE, ip,
 		    (caddr_t)ctob(sptov(u.u_procp, u.u_ssize - 1)),
-		    ctob(u.u_ssize),
-		    ctob(UPAGES)+ctob(u.u_dsize), 0, (int *)0);
+		    (int)ctob(u.u_ssize),
+		    (off_t)ctob(UPAGES)+ctob(u.u_dsize), 0, (int *)0);
 out:
 	iput(ip);
 	return (u.u_error == 0);
